Add book module to isaolate create book responsability
