+++ book_module.rb
@@ -1,0 +1,12 @@
+module BookModule
+  # Create a book.
+  def create_book
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    @books.push(Book.new(title, author))
+    puts 'Book created successfully'
+    @option.show_options
+  end
+end
